requirements.txt
# Core AWS SDK
boto3==1.34.50
botocore==1.34.50

# AWS CDK (for infrastructure)
aws-cdk-lib==2.120.0
constructs>=10.0.0

# AI/ML
langchain==0.1.6
langchain-aws==0.1.0

# Web Framework
streamlit==1.31.0
streamlit-extras==0.3.6

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-multipart==0.0.6

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.4
moto==5.0.0
faker==22.6.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1

# Logging
loguru==0.7.2

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0